feat: Update Client-Management and QuoteEstimate components
This commit updates the Client-Management component by fixing a typo in the map function parameter name. It changes `clients.map((Client) => (` to `clients.map((client) => (`.

It also adds a new component called QuoteEstimate. The QuoteEstimate component is responsible for displaying an estimate for a quote. It takes a prop called `estimate` which is a number representing the estimated price. The component renders the estimate with two decimal places and includes a disclaimer about the estimate being subject to change.

These changes improve the functionality and user experience of the application.